% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{clean_date}
\alias{clean_date}
\title{Add time to date variable}
\usage{
clean_date(data, startdate, enddate)
}
\arguments{
\item{data}{A data frame.}

\item{startdate}{Start date variable.}

\item{enddate}{End date variable.}
}
\value{
data.frame
}
\description{
Add time to date variable
}
