% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{cox_fit}
\alias{cox_fit}
\title{Fit Cox Regression - wrapper for \code{survival::coxph()}}
\usage{
cox_fit(data, splits)
}
\arguments{
\item{data}{A data frame.}

\item{splits}{Variables to use in Cox regression}
}
\description{
Fit Cox Regression - wrapper for \code{survival::coxph()}
}
