% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{create_model}
\alias{create_model}
\title{Create Weibull Model and Plot of Data}
\usage{
create_model(data)
}
\arguments{
\item{data}{A data frame}
}
\value{
Interactive plotly plot including KM curve of data and Weibull model of data
}
\description{
Create Weibull Model and Plot of Data
}
\details{
wrapper for \code{survminer::ggsurvplot()} and \code{survival::survreg()}
}
