% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{create_SurvPlot}
\alias{create_SurvPlot}
\title{Create KM curves: wrapper for \code{survival::survfit} and \code{survminer::ggsurvplot}}
\usage{
create_SurvPlot(data, splits)
}
\arguments{
\item{data}{A data frame.}

\item{splits}{variables used to create km plot strata}
}
\value{
prints a km plot plot based on user entered strata
}
\description{
Create KM curves: wrapper for \code{survival::survfit} and \code{survminer::ggsurvplot}
}
