% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{lt_cols}
\alias{lt_cols}
\title{Return Searched colors}
\usage{
lt_cols(...)
}
\arguments{
\item{...}{A list specifying the colors from lt_colors to return}
}
\value{
A list of colors to use with \code{ggplot}
}
\description{
Return Searched colors
}
\examples{
lt_cols("Twinkle Twinkle", "Electric Salmon", "Fashion Fuschia")
}
