% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{lt_pal}
\alias{lt_pal}
\title{Create a color ramp}
\usage{
lt_pal(palette = "main", reverse = FALSE, ...)
}
\arguments{
\item{palette}{A pallet selected from lt_palettes which itself is composed of lt_colors}

\item{reverse}{Binary argument that reverses the palette order}

\item{...}{selected color list}
}
\value{
\code{colorRampPalette()} of selected colors
}
\description{
Create a color ramp
}
