% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{return_split}
\alias{return_split}
\title{Extract Piece of Date - yyyy, mm, dd}
\usage{
return_split(date, return_num)
}
\arguments{
\item{date}{A POSIXct, Date, or string representing a date.}

\item{return_num}{The piece (separated by - or /) to return.}
}
\value{
A character string representing the selected piece of the date.
}
\description{
Extract Piece of Date - yyyy, mm, dd
}
\examples{
return_split(date = "2021-11-30", return_num = 2)
}
