% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{scale_color_lt}
\alias{scale_color_lt}
\title{Create a scale for \code{color}}
\usage{
scale_color_lt(palette = "main", discrete = TRUE, reverse = FALSE, ...)
}
\arguments{
\item{palette}{A pallete selected from lt_palettes}

\item{discrete}{A binary argument specifying if colored values are discrete}

\item{reverse}{A binary argyment that reverses the color palette}

\item{...}{color list used to create gradient}
}
\value{
a color palette to be used in ggplot
}
\description{
Create a scale for \code{color}
}
