% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{server}
\alias{server}
\title{Shiny app server function}
\usage{
server(input, output, session)
}
\arguments{
\item{input}{Provided by shiny.}

\item{output}{Provided by shiny.}

\item{session}{Provided by shiny}
}
\description{
Shiny app server function
}
